pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ergane"
version = "0.2.0"
description = "High-performance async web scraper with selectolax parsing"
authors = [{name = "pyamin1878"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
keywords = ["web-scraper", "crawler", "async", "httpx", "selectolax", "spider"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Framework :: AsyncIO",
]
dependencies = [
    "httpx[http2]>=0.27.0",
    "selectolax>=0.3.21",
    "polars>=1.0.0",
    "pydantic>=2.0.0",
    "click>=8.1.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "mypy>=1.8.0",
    "ruff>=0.2.0",
]

[project.urls]
Homepage = "https://github.com/pyamin1878/ergane"
Repository = "https://github.com/pyamin1878/ergane"
Issues = "https://github.com/pyamin1878/ergane/issues"

[project.scripts]
ergane = "src.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
